Updated Dockerfile and start_textpresso.sh for better initialization
The start of the cron job was moved to the bash script 'start_textpresso.sh' from Dockerfile to ensure the cron job begins whenever the script is run. Also, two directories were created in the script for a more efficient data organization. This change improves the initialization process of textpresso API services and ensures the cron job is always started properly with the script.