Refactor incremental build script and update .gitignore
Simplified directory removal process in the incremental_build.sh script resulting in cleaner, more concise code. Instead of checking if a directory exists before moving it to an .old version, it now directly removes the directories. This change was made to improve script performance as checking for directory existence and then moving it was redundant.

.gitignore was also updated to include pattern-based exclusions for raw_files and tpcas-* folders. This allows us to better filter out unnecessary build and execution artifact folders from git tracking. We adopted this exclusion strategy to keep our repository clean and focused only on the source code.