Change depth of image search in `03pdf2cas4tai.sh`
The depth of the find operation for copying images was changed from 2 to 1 in `03pdf2cas4tai.sh`. This alteration avoids unnecessary directory traversal, boosting script efficiency. Previously, images in subdirectories of the source directory were unnecessarily traversed and copied, a case not needed in the given context, resulting in wasteful operations.